isInList() function bug fix, addToCart() function done
in the isInList() function the bug was to: checking from the listOfProducts list, not from the given list. this is fixed. 
addToCart() function if fully implemented. while product is not in the cart adding a new object in the list, if item is already in the list, just increasing the quantity.